+++ src/core/index.ts
@@ -1,10 +1,10 @@
 // 🌲☀️🌲
 // This file is the entry point for the core utilities
 // that have no dependencies on external libraries.
-export { getFrameHtmlResponse } from '../frame/getFrameHtmlResponse';
-export { getFrameMetadata } from '../frame/getFrameMetadata';
-export { getFrameMessage } from '../frame/getFrameMessage';
-export { getMockFrameRequest } from '../frame/getMockFrameRequest';
+export { getFrameHtmlResponse } from '../frame/utils/getFrameHtmlResponse';
+export { getFrameMetadata } from '../frame/utils/getFrameMetadata';
+export { getFrameMessage } from '../frame/utils/getFrameMessage';
+export { getMockFrameRequest } from '../frame/utils/getMockFrameRequest';
 export type {
   FrameButtonMetadata,
   FrameImageMetadata,
