+++ src/frame/index.ts
@@ -2,10 +2,10 @@
 // This file is the entry point for the core utilities
 // that have no dependencies on external libraries.
 export { FrameMetadata } from './components/FrameMetadata';
-export { getFrameHtmlResponse } from './getFrameHtmlResponse';
-export { getFrameMetadata } from './getFrameMetadata';
-export { getFrameMessage } from './getFrameMessage';
-export { getMockFrameRequest } from './getMockFrameRequest';
+export { getFrameHtmlResponse } from './utils/getFrameHtmlResponse';
+export { getFrameMetadata } from './utils/getFrameMetadata';
+export { getFrameMessage } from './utils/getFrameMessage';
+export { getMockFrameRequest } from './utils/getMockFrameRequest';
 export type {
   FrameButtonMetadata,
   FrameData,
