+++ src/frame/utils/getMockFrameRequest.ts
@@ -2,7 +2,7 @@ import type {
   FrameRequest,
   MockFrameRequest,
   MockFrameRequestOptions,
-} from './types';
+} from '../types';
 
 /**
  * Modify a standard frame request to include simulated values (e.g., indicate the viewer
