+++ src/network/client.ts
@@ -1,4 +1,4 @@
-import { createPublicClient, http } from 'viem';
+import { http, createPublicClient } from 'viem';
 import { mainnet } from 'viem/chains';
 
 export const publicClient: ReturnType<typeof createPublicClient> =
