+++ src/network/getChainPublicClient.ts
@@ -1,4 +1,4 @@
-import { createPublicClient, http } from 'viem';
+import { http, createPublicClient } from 'viem';
 import type { Chain } from 'viem/chains';
 
 export function getChainPublicClient(chain: Chain) {
