+++ src/network/neynar/getDataFromNeynar.ts
@@ -1,6 +1,6 @@
 import { version } from '../../version';
-import { FetchError } from './exceptions/FetchError';
-import { NEYNAR_DEFAULT_API_KEY } from './frame/neynarFrameValidation';
+import { FetchError } from './FetchError';
+import { NEYNAR_DEFAULT_API_KEY } from './neynarFrameValidation';
 
 export async function getDataFromNeynar(
   url: string,
