+++ src/network/neynar/neynarBulkUserLookup.ts
@@ -1,5 +1,5 @@
-import { getDataFromNeynar } from '../getDataFromNeynar';
 import { convertToNeynarResponseModel } from './convertToNeynarResponseModel';
+import { getDataFromNeynar } from './getDataFromNeynar';
 import type { NeynarBulkUserLookupModel } from './types';
 
 export const NEYNAR_DEFAULT_API_KEY = 'NEYNAR_ONCHAIN_KIT';
