+++ src/network/neynar/neynarFrameValidation.ts
@@ -1,6 +1,6 @@
-import type { FrameValidationData } from '../../../frame/types';
+import type { FrameValidationData } from '../../frame/types';
 import { convertToNeynarResponseModel } from './convertToNeynarResponseModel';
-import { postDataToNeynar } from '../postDataToNeynar';
+import { postDataToNeynar } from './postDataToNeynar';
 
 export const NEYNAR_DEFAULT_API_KEY = 'NEYNAR_ONCHAIN_KIT';
