+++ src/network/neynar/types.ts
@@ -1,3 +1,22 @@
+export interface NeynarUserModel {
+  fid: number;
+  custody_address: string;
+  username: string;
+  display_name: string;
+  pfp_url: string;
+  profile: {
+    bio: {
+      text: string;
+    };
+  };
+  follower_count: number;
+  verifications: string[];
+}
+
+export interface NeynarBulkUserLookupModel {
+  users: NeynarUserModel[];
+}
+
 /**
  * Raw Response from Neynar
  */
